requirements.txt
# Python包依赖
# 在此文件中添加项目所需的Python包
# 格式: package_name==version

# Flask web框架
flask==3.0.0
flask-cors==4.0.0

# Google AI SDK
google-genai==1.29.0
google-cloud-aiplatform==1.56.0

# 其他工具
python-dotenv==1.0.0
requests==2.31.0 